requirements.txt
h5py==3.1.0
numpy==1.19.5
Pillow==8.2.0
torch==1.1.0
torchvision==0.3.0
tqdm